fexed segfault when removing 2nd modulation

+++ Audio/Modulators/Perlin.h
@@ -37,7 +37,7 @@ public:
     }
     float getNextValue()
     {
-        currentX += (currentFreq / (AudioSystem::getSampleRate() * 6.0f));
+        currentX += (currentFreq / ((float)AudioSystem::getSampleRate() * 6.0f));
         return Perlin::getFractal(currentX, currentOctaves, currentFreq, currentLacunarity);
     }
